+++ Assets/EVE/Scripts/Utils/GameObjectUtils.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using System.Linq;
 using UnityEngine;
 
 namespace Assets.EVE.Scripts.Utils
@@ -11,7 +12,7 @@ namespace Assets.EVE.Scripts.Utils
         /// <param name="gameObject">To be destroyed</param>
         /// <param name="delay">Time to destruction</param>
         /// <returns>Coroutine to execute the delay</returns>
-        public static IEnumerator RemoveGameObject(GameObject gameObject, int delay = 0)
+        public static IEnumerator RemoveGameObject(GameObject gameObject, float delay = 0)
         {
             yield return new WaitForSeconds(delay);
             Object.Destroy(gameObject);
@@ -27,6 +28,31 @@ namespace Assets.EVE.Scripts.Utils
             var newObject = Object.Instantiate(Resources.Load(prefab)) as GameObject;
             if (newObject == null) Debug.LogError("Failed instantiating " + prefab);
             return newObject;
+        }
+
+
+        /// <summary>
+        /// Find a named game object with depth first search.
+        /// </summary>
+        /// <remarks>
+        /// Note that only the first game object of the name will be returned.
+        /// </remarks>
+        /// <param name="container">Root of hierarchy to be searched.</param>
+        /// <param name="gameObjectName">Name of game object to be found.</param>
+        /// <returns>First instance of a game object with requested name.</returns>
+        public static Transform FindGameObjectInChildren(Transform container, string gameObjectName)
+        {
+            Transform dynf = null;
+            var allGameObjects = container.GetComponentsInChildren<Transform>().ToList();
+            foreach (var child in allGameObjects)
+            {
+                if (child.name == gameObjectName)
+                {
+                    dynf = child;
+                    break;
+                }
+            }
+            return dynf;
         }
     }
 }